+++ src/domain/body/obstacle/SimpleObstacle.java
@@ -2,12 +2,12 @@ package domain.body.obstacle;
 import domain.needForSpear.*;
 
 public class SimpleObstacle extends Obstacle {
-    private double vx;
+    private int vx;
     private String name;
-    public SimpleObstacle(double x_coordinates,
-                          double y_coordinates,
-                          double length,
-                          double width,
+    public SimpleObstacle(int x_coordinates,
+                          int y_coordinates,
+                          int length,
+                          int width,
                           int numberOfHits) {
         super(x_coordinates, y_coordinates, length, width, numberOfHits);
         vx = Controller.getInstance().getPlayer().getNoblePhantasm().width/(4*Controller.ticksPerSecond);
